export pdf function and remove direct dom manipulation
