+++ met-web/src/components/publicDashboard/SurveyBar.tsx/index.tsx
@@ -63,54 +63,56 @@ export const SurveyBar = ({ engagement, engagementIsLoading }: SurveyQuestionPro
     }
 
     if (isError) {
-        <>
-            <Grid item xs={12}>
-                <MetLabel mb={2} color="primary">
-                    Survey Results
-                </MetLabel>
-            </Grid>
-            <Grid ml={isSmallScreen ? 0 : 5} item xs={12}>
-                <MetPaper sx={{ p: 2 }}>
-                    <Grid item xs={12}>
-                        <Stack direction={{ xs: 'column', sm: 'row' }} width="100%" justifyContent="flex-end">
-                            <Grid item container xs={12} direction="row" justifyContent="center">
-                                <MetLabel mb={2} color="primary">
-                                    Click on a question to view results
-                                </MetLabel>
-                                <ToggleButtonGroup value={chartType} exclusive onChange={handleToggleChange}>
-                                    <ToggleButton
-                                        value="bar"
-                                        sx={{
-                                            '&.Mui-selected': {
-                                                backgroundColor: Palette.primary.main,
-                                                color: 'white',
-                                            },
-                                        }}
-                                    >
-                                        Show as Bar Chart
-                                    </ToggleButton>
-                                    <ToggleButton
-                                        value="treemap"
-                                        sx={{
-                                            '&.Mui-selected': {
-                                                backgroundColor: Palette.primary.main,
-                                                color: 'white',
-                                            },
-                                        }}
-                                    >
-                                        Show as TreeMap Chart
-                                    </ToggleButton>
-                                </ToggleButtonGroup>
+        return (
+            <>
+                <Grid item xs={12}>
+                    <MetLabel mb={2} color="primary">
+                        Survey Results
+                    </MetLabel>
+                </Grid>
+                <Grid ml={isSmallScreen ? 0 : 5} item xs={12}>
+                    <MetPaper sx={{ p: 2 }}>
+                        <Grid item xs={12}>
+                            <Stack direction={{ xs: 'column', sm: 'row' }} width="100%" justifyContent="flex-end">
+                                <Grid item container xs={12} direction="row" justifyContent="center">
+                                    <MetLabel mb={2} color="primary">
+                                        Click on a question to view results
+                                    </MetLabel>
+                                    <ToggleButtonGroup value={chartType} exclusive onChange={handleToggleChange}>
+                                        <ToggleButton
+                                            value="bar"
+                                            sx={{
+                                                '&.Mui-selected': {
+                                                    backgroundColor: Palette.primary.main,
+                                                    color: 'white',
+                                                },
+                                            }}
+                                        >
+                                            Show as Bar Chart
+                                        </ToggleButton>
+                                        <ToggleButton
+                                            value="treemap"
+                                            sx={{
+                                                '&.Mui-selected': {
+                                                    backgroundColor: Palette.primary.main,
+                                                    color: 'white',
+                                                },
+                                            }}
+                                        >
+                                            Show as TreeMap Chart
+                                        </ToggleButton>
+                                    </ToggleButtonGroup>
+                                </Grid>
+                            </Stack>
+                            <Divider sx={{ marginTop: '1em' }} />
+                            <Grid container direction="row" item xs={12} spacing={1} alignItems={'flex-start'}>
+                                <CircularProgress color="inherit" />
                             </Grid>
-                        </Stack>
-                        <Divider sx={{ marginTop: '1em' }} />
-                        <Grid container direction="row" item xs={12} spacing={1} alignItems={'flex-start'}>
-                            <CircularProgress color="inherit" />
                         </Grid>
-                    </Grid>
-                </MetPaper>
-            </Grid>
-        </>;
+                    </MetPaper>
+                </Grid>
+            </>
+        );
     }
 
     return (
